gom: Don't use GSimpleAsyncResult in worker thread
Using GSimpleAsyncResults in the worker thread makes it
impossible to chain calls inside the SQLite thread, making
it hard to implement reliable synchronous calls.

Instead, we'll use our own structure, so that the callbacks
passed to the worker thread will be responsible for sending
out results through GSimpleAsyncResult if they want to use
such a construct.