gom: Enhance performance for our use cases
First, we'll use the SQLITE_OPEN_NOMUTEX open flag, which will
mean that we are expected to implement our own locking access to
the database, which we don't need to do as a single worked thread
will access the database.

Secondly, we'll execute a few PRAGMA statements. The journal_mode
will change the behaviour of the rollback model. We now use a
write-ahead log instead of a journal. This is much faster with small
transactions (ours are very small).
https://www.sqlite.org/pragma.html#pragma_journal_mode

With this change, adding 1000 items into a disk-backed database
drops from 20 seconds to 7 seconds.

Finally, we also change the synchronous behaviour to NORMAL.
This is a good trade-off between performance, and possibility of
database corruption.
https://www.sqlite.org/pragma.html#pragma_synchronous

With the same test case as above, we drop from taking 7 seconds
to 0.2 seconds to add 1000 items.