all: Stop first-time GOM users from having to set the "new in version"
Applications that use GOM from the get-go shouldn't have to
call "new in version" for every single one for their initial
properties/columns.

Nor should applications with existing columns have to set their
"new in version" to 0 for those columns.

Instead, we'll try to create the table with all the columns
that are new in version 1 in one go, which can fail gracefully.