gom: Add GomError enum as a single unified error domain
Adds GomError as a single error domain replacing GomCommandError,
GomResourceError etc. The enumerators are taken from the existing
four error domains used in Gom.

This also adds a gom_error_get_type() function to define a GType
for GomError, which is necessary for introspection purposes. It
might be better to use glib-mkenums for this purpose.

https://bugzilla.gnome.org/show_bug.cgi?id=728491