tests: Add failing test with non-automatic primary keys
When the primary key value of a GomResource is set, we imagine that it
must mean that the value is already present in the database. But for
non-integer primary keys, the caller is responsible for generating the
primary key, so it *must* be set before we call gom_resource_save*.

This means that we'll silently fail to insert the item in the database.