gom: Fix another memory management problem in GomFilter
Follow-up to 93694570fde8e29dff132644baa2efb18ef31ff5.

When appended to the target array, the values aren't refcounted, so
destroying the whole source array would destroy the values as well.
Disable the freeing of the individual items in the source array so
that they're not destroyed along with the source array. The destination
array will now assume ownership.