sorting: Fix the enum nicks
When I wrote the sorting patches, I was at first using the nick strings
as the string to insert in the SQL. As a result, I used "DESC" for the
descending mode, and "" for the ascending mode, as for an ascending
ORDER BY, you just don't add anything to the SQL.

But then I realize it wasn't such a great idea to use the nick strings
in this way, so the versions of the sorting patches that got merged
don't do that any more.

However I kept those nick strings in the enum.

Turns out that GI-based bindings use the nick as the identifier for the
enum member.

For example, in Python, the descending mode would be:

    Gom.SortingMode.DESC

Obviously, the empty string is not a great identifier.

This commit fixes that, giving them proper nicks, which make things
nicer to use in GI-based bindings.