gom: Add support for creating new tables in a version
To support creating new tables in a particular version of the database,
we special case the primary key's "new in version" tag to create the
table in that version.

For example, if a table with 2 columns, one primary key, and one
attribute, needs to be created in version 4 of the database, you'll need
to tag both the primary key property as well as the attribute one as
being new in version 4 of the database.

Think of this as an easter egg, as you'd really be expected to create
all the necessary tables in version 1, or to manually migrate your
database otherwise.

https://bugzilla.gnome.org/show_bug.cgi?id=753563