gom: Store NULL GDateTimes in the database as NULL
Currently Gom will store NULL GDateTimes in the database as the epoch,
1970-01-01T00:00:00Z, instead of NULL.  This unneccessarily increases
the storage requirements of every GDateTime column in every row by 20
bytes.  This also presents a loss of information: there's no way to
distinguish between the valid input values of the epoch and NULL.  The
API user has to account for the way in which the library is
transforming data values.

To fix this, we store NULL GDateTime values in the database as NULL
instead.