resource-group: fix concurrent fetch and get_index
If fetch was active while a get_index() was called, the GHashTable could
be mutated while a reader exists.

This fixes that by protecting it with a mutex and then inserting all of
the items from the worker in one go.

Such use is what you might expect if you're paginating data and lazy
loading regions of the resource group.