gom/resource-group: avoid hashtables for itemdata
Instead use a union that lets us switch bewteen the GParameter array and
the item resource. Keep track of how many parameters we have in there too.

Optimize the inflation of groups by finding the mappings once and then
using that instead of column iteration. Avoid a lot of GTypeClass ref'ing
as well. Only look up the quark data once per mapped parameter too.

We keep an extra GParameter at the end so that we can avoid copying the
GParameter for inflation somewhere else. We just add the repository at
the end of it before constructing/replacing the inflated instance.